// src/app/core/services/model/controllerUpdateError.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type ControllerUpdateError = 'UpdateNotStartedYet' | 'UpdateRunning' | 'MockUpdateSuccessfulWindows' | 'UpdateSuccessfulRebooting' | 'UploadFileFailed' | 'UpdateFailedUnknownReason' | 'RaucInfoFailed' | 'UpdateFileProductIdWrongFormat' | 'CannotReadControllerSerialNumber' | 'ProductIdIsNotMatching' | 'UnableToGetLocalIssuerCertifcate' | 'UnableToGetIssuerCertificate' | 'InvalidCertificate' | 'UpdatedAlreadyInstalled' | 'UpdateIsOlderThanCurrentVersion' | 'DecryptingUpdateFileFailed';

export const ControllerUpdateError = {
    UpdateNotStartedYet: 'UpdateNotStartedYet' as ControllerUpdateError,
    UpdateRunning: 'UpdateRunning' as ControllerUpdateError,
    MockUpdateSuccessfulWindows: 'MockUpdateSuccessfulWindows' as ControllerUpdateError,
    UpdateSuccessfulRebooting: 'UpdateSuccessfulRebooting' as ControllerUpdateError,
    UploadFileFailed: 'UploadFileFailed' as ControllerUpdateError,
    UpdateFailedUnknownReason: 'UpdateFailedUnknownReason' as ControllerUpdateError,
    RaucInfoFailed: 'RaucInfoFailed' as ControllerUpdateError,
    UpdateFileProductIdWrongFormat: 'UpdateFileProductIdWrongFormat' as ControllerUpdateError,
    CannotReadControllerSerialNumber: 'CannotReadControllerSerialNumber' as ControllerUpdateError,
    ProductIdIsNotMatching: 'ProductIdIsNotMatching' as ControllerUpdateError,
    UnableToGetLocalIssuerCertifcate: 'UnableToGetLocalIssuerCertifcate' as ControllerUpdateError,
    UnableToGetIssuerCertificate: 'UnableToGetIssuerCertificate' as ControllerUpdateError,
    InvalidCertificate: 'InvalidCertificate' as ControllerUpdateError,
    UpdatedAlreadyInstalled: 'UpdatedAlreadyInstalled' as ControllerUpdateError,
    UpdateIsOlderThanCurrentVersion: 'UpdateIsOlderThanCurrentVersion' as ControllerUpdateError,
    DecryptingUpdateFileFailed: 'DecryptingUpdateFileFailed' as ControllerUpdateError
};