// src/app/core/services/model/controllerUpdateStatus.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ControllerUpdateError } from './controllerUpdateError';

export interface ControllerUpdateStatus { 
    controllerUpdateError?: ControllerUpdateError;
    /**
     * e.g. StandardErrorOutput
     */
    additionalErrorInformation?: string;
}